[WebAPKs] Refactor the PWA install bottom sheet
Refactor the PWA install bottom sheet to make it more dynamic and
scalable with size changes.
Merged the divider in the content as it wasn't aware of the toolbar and
that caused overlapping issue with font scaling at 200%.

Content layout should be refactored to use ConstraintLayout after we
remove the WebLayer from WebView.


Bug: 1422063
Change-Id: If2b57f1bd3fef3239ac78715da5aaaadb582b405
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4580219
Reviewed-by: Glenn Hartmann <hartmanng@chromium.org>
Commit-Queue: Sayed Elabady <elabadysayed@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157557}