[iOS] Autofill branding: implement behaviors for variations
This CL implements variations defined in ios/c/b/flags/about_flags.mm
plus the default enabled (two impressions) and the disabled arms.

Variation 1: Show two impressions on both iPhone and iPad
Variation 2: Show the branding until the user has interacted with a
form input accessory on both iPhone and iPad
Variation 3: Always show the branding on iPad; on iPhone, always show
the branding and make it slide away from the leading edge
Variation 4: Always show the branding on iPad; on iPhone, show the
branding until the user has interacted with a form input accessory, and
when the user does so, the branding should slide away from the leading
edge

Test case for each variation is also implemented.

As of now, the code and test for the "slide away from leading edge"
animation has not been implemented yet. This will be done in a follow up
CL.

Bug: 1447909
Change-Id: I3bf626c3cfcc94b716af559391856188e5e8a8cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4605159
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1157566}