[ios] Simplify code that handles height-above-feed changes
This changes the method that gets called when the Doodle finishes
loading asynchronously. Now instead of updating the height-above-
feed and scrolling to the top, it will update the height-above-
feed and keep the current scroll position, just like we do for
MVT and other content suggestion updates.

Change-Id: If185c556891bdfb9d5ca4bb5e47c72f553a333b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4605126
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1157592}