Add a missing call to ComputeEntirelyCoveredByBucketing().
RuleSets that were inserted into non-keyed buckets did not get a
proper call to ComputeEntirelyCoveredByBucketing(). This also meant
that they would never be marked as easy, which meant that selectors
like “a:-webkit-any-link” would not be processed as easy.

Style recalc is only 0.4% faster, which shows that this was not a
big issue, but it's good for completeness.

Change-Id: I83d157448533cb86429aaa70a22c99b824b7cfb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4604628
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Auto-Submit: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157643}