[X-Device Restore] Add thumbnail placeholder icon
For tabs that do not have a thumbnail show a placeholder icon.

This is behind a flag ThumbnailPlaceholder and will be rolled out along
with the Cross Device Tab Restore flag. Which is being developed
as a separate component. Given the tabs for that feature will be
restored without thumbnails it is important to convey this is
intentional whereas the current placeholder color is possibly indicative
of loading or some other error has occurred.

This CL also adds
* Render tests for the new and old behaviors.
* Cleans up some layout defaults which is possible after changing phones
  to use GRID_CARD_SIZE in an earlier CL.

Mock images:
https://drive.google.com/file/d/1KGpea8jzrhRaqnuk751C2oHGyWcGrznu/view?usp=sharing

Bug: 1441433
Change-Id: I709a002e207462c0b12893230ff55d158c4c3304
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4456147
Reviewed-by: Mei Liang <meiliang@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1157690}