lacros/server: Propagate top inset from lacros to ash
Top inset property is used in ash to clip away header in overview and
mirror views and others. This is the server side implementation for
lacros windows to set their associated exo window's top inset.

ClientControlledShellSurface provides the logic for setting the top
inset on an exo window, so move it to ShellSurfaceBase where it can be
used by other exo windows in the future.

Bug: b/286417311
Test: exo_unittests *SetTopInset*
Change-Id: I4610125942756db54ee13eac033c56ed5457a275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4609224
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157691}