[omnibox][CR23] Update suggestion hover fill shape per CR23 spec.
This CL updates the suggestion hover fill shape based on the latest
ChromeRefresh2023 guidance. More specifically, the shape of the
background rendered for suggestions in the omnibox popup is now rounded
on the right-hand side and shifted in by 16px.

One point to note is that the newly-introduced 16px space on the
right-hand side is a "dead zone" which does not react to user input
(click/hover) at this time. The exact bounds of this "dead zone" are
indicated by the second row of
https://screenshot.googleplex.com/63bJEFmco78Qsb8. In particular, the
area noted in blue reacts to user input, while the area noted in red
does not.

In order to support experimentation with this feature, a new
OmniboxSuggestionHoverFillShape feature flag is also being introduced by
the changes in this CL.

Low-Coverage-Reason: ChromeRefresh2023 UI update needs to be  experimented with prior to investing in test coverage.
Bug: 1431337
Change-Id: I76fdde3b0d83e05cd97c0816a1eeb65f08412f2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4611098
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Khalid Peer <khalidpeer@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157699}