immersive_mode: Set background to the top container view
This CL sets the background with the dynamic color to the top container
view when it's revealed in the immersive mode. It also avoid calling
BrowserNonClientFrameViewChromeOS::Paint to paint the frame header,
since it will be triggered by BrowserRootView::PaintChildren(). We can
avoid calling it twice.

Before fix:
 1. Light Mode: https://screenshot.googleplex.com/9WLYoKjRuX4wuhv
 2. Dark Mode: https://screenshot.googleplex.com/gS8pYWi5AeVshTo

After fix:
 1. Light Mode: https://screenshot.googleplex.com/6bX6cpjunrFrcMy
 2. Dark Mode: https://screenshot.googleplex.com/7Toa9xMJkH6Hzf3

Bug: b/277943803
Change-Id: Ia7fd7c3249fcef65e6f585fb58231bd55b9e270b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4591042
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Connie Xu <conniekxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157710}