Add mandatory re-auth manager backend component (3/4)
This CL adds the mandatory re-auth manager backend component for the re-
auth opt-in flow. It is CL 3/4, and in the final CL we will link
everything together to trigger the re-auth opt-in flow.

The authentication flow will later be refactored into the
MandatoryReauthManager as well.

test_autofill_client.h, which is just test code. It is also possible
the coverage delta is not accurately reported in the case of inlined
code in .h files, as this code is being used in tests. I have filed
crbug.com/1448697 to double check this.

Bug: 4555994
Change-Id: I9433518bd7e1d8662764d4e9b912756cde60eb61
Low-Coverage-Reason: The low coverage warning is for
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4556577
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Siyu An <siyua@chromium.org>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Reviewed-by: Dominic Battre <battre@chromium.org>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Cr-Commit-Position: refs/heads/main@{#1157726}