[PDF Viewer] Validate params page value input is in bounds
When parsing the URL param `page`, the page number provided by the user
may be out of bounds of the number of pages. The page number should be
clipped when out of bounds.

To check the upper bound, OpenPdfParamsParser must store the page count.

Fixed: 1453872
Change-Id: I097c821f537526621c83274b4d49250aed643151
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4608385
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157779}