Add discard and memory usage accessible strings
Adds tab discard and memory usage accessible strings so when the user
tabs over a discarded tab, or a tab that has memory usage information,
the screen reader will announce that the tab is discarded with the
amount of memory saved through discarding or the tab's current memory
usage.

Note that accessible labels for tab alerts are usually more succinct
than what is written on the hover cards. So the string that a screen
reader will read aloud differs slightly from what is written on the
hover card.

Bug: 1424226
Change-Id: I8fbf15dbdc70eedcd8b2580dfb1b1dec0f87c03b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4601817
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157803}