[SparseObjectPaintProperties #⁠2] Implement pure virtual ObjectPaintProperties
This patch is the second in a series for the SparseObjectPaintProperties
experiment. In this patch, the ObjectPaintProperties is changed from an
empty class to a pure virtual base class for ObjectPaintPropertiesImpl.

NOTE: after the conclusion of the SparseObjectPaintProperties
experiment, this class will be deleted--so to keep it similar to how
the ObjectPaintPropertiesImpl class is implemented, it uses the same
style of preprocessor macros, even though they are something of an
anti-pattern.

The Project:
SparseObjectPaintProperties is a cleanup project with the goal of
lowering the total memory cost of the ObjectPaintProperties class,
through the use of a new SparseVector class that allows for zero
memory overhead for null pointers. If successful, this project should
significantly the total memory usage of this class and hopefully
impact Chrome's memory usage in general.

Bug: 1454638
Change-Id: I643c0aa6b5a3e593379f8d738ae7aa450164221f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4611813
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157804}