[Android] Obvious error when clicking a view that hasn't been sized
This is a common source of flakiness, but the error message is usually
non-descriptive because a click on pixel (0, 0) is legal. We never want
to click a view that hasn't been sized yet and this surfaces the error
earlier and with a better message.

Bug: 1444083,1441822,1435426,1444217,1444241,1444222
Change-Id: If1dff0ebb2f52218e25dd15be1c20ddff538daa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4608982
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157844}