Refactor BluetoothBlueZTest to fix flaky discovery test - part 1
This is a follow-up CL to CL4568899. The original test was flaky due to
a race condition that would occasionally cause the run loop to break
before the expected task of updating the fake adapter client properties
could be executed. This CL refactors the code to ensure that we are validating internal states at the right moment.

Bug: 1446496
Change-Id: I7338cf9e763e66d720e9800b1a0b90e2408e20c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4595414
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Alvin Ji <alvinji@chromium.org>
Commit-Queue: Alvin Ji <alvinji@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157867}