Wayland protocol to allow exo to rotate view focus in lacros - Client
When focus rotation is requested for lacros widget, a rotate focus event
is emitted from exo to lacros. These requests are fenced with a serial,
which will prevent delay in processing from triggering multiple
subsequent rotations in ash.

When the client has reached the end of it focusable views, it will
notify the the exo with the aforementioned serial.

For this change, much of the rotate pane focus code had to be moved to
the widget delegate. This is because the behaviour is now platform
dependent.

Change-Id: Ifbe31230578444cc35db8143fc5cfd0c14448b95
Bug: 1361120
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4554431
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1157882}