drm: Only use modifiers that support Vulkan import
Currently when Vulkan is enabled, we do not take into account whether
a particular DRM format modifier is supported for Vulkan import via
vkCreateImage. Thus, we can allocate a presentation buffer with a
modifier that then cannot be imported into Vulkan, and later we fail
display composition. For example, the Intel Vulkan drivers do not
support import of Gen 12 CCS buffers, and Chrome crashes on startup
if we enable Vulkan on a sufficiently modern Intel GPU.

This CL fixes the issue by querying for supported modifiers when
populating the available modifiers list and only returning those that
can be imported.

Change-Id: Iaa9ca90f97caddf0968489400aea745780a45257
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4262572
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Brian Ho <hob@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157885}