[FilesBulkPinning] Fix up failing test before enabling Progress events
The OnSyncingStatusUpdate_SyncStatusTracksStatus test fails when we
enable the kFilesInlineSyncStatusProgressEvents as it expects that flag
to be false, explicitly override the flags in the test to ensure
assertion of the correct flow. This test can be removed when we remove
the flags.

Bug: b:272585066
Test: chromeos_unittest --gtest_filter=*SyncStatusTracksStatus*
Change-Id: Ib47cb6a7a1a7dd0bd84fe93bc3057ebfffd5c423
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4615465
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Auto-Submit: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157902}