Update anchor bounds for ShelfView.
The ShelfView bounds fill available space in the shelf, but the user
perceived bounds (where help bubbles should be anchored to) are
aligned with the center of the display and visible items in the shelf.

This CL updates anchor bounds to align with user perceived bounds.

Before: http://shortn/_PcKDXyZUsg
After:  http://shortn/_uDNTfpUpIs

Bug: b:277091650
Change-Id: Id3f820f1a74e4cc60d921926170e29eae7a6a5c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4611732
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/main@{#1157933}