Integrate backend components for mandatory re-auth opt-in (4/4)
This CL integrates the backend components for the mandatory re-auth
opt-in flow, so that every time there is a green path autofill on a
form, we prompt re-auth upon form submission if the user should be
prompted for re-auth. It is the last CL in the chain for the re-auth
opt-in flow.

Bug: 1427216
Change-Id: I4b1f2063beb3cd2dd2a8b70191905fff72482028
Low-Coverage-Reason: The low coverage warning is for test_autofill_client.h, which is strictly a class used for testing. I've also filed crbug.com/1448697 as I believe there might be a bug in the low coverage check for inlined functions in the .h file.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4582515
Reviewed-by: Dominic Battre <battre@chromium.org>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157948}