Fix nullptr issue with in session optin
After user accept ARC TOS in in-session optin dialog, we update the UI
asynchronously to loading page. But when system is busy it is possible
that the UI update happens late and user clicks the accept button again.

Ignore the late click to avoid calling on nullptr.

Bug: b/284000632
Test: Trybot
Change-Id: I11fb03da3139075a3c6f3854b88c0bd6213bd509
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4616396
Commit-Queue: Long Cheng <lgcheng@google.com>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1157952}