Enable ExtensionControlHandler in the settings app
This CL fixes an issue where user-installed proxy extensions cannot be
disabled by the disable button in the ChromeOS settings app.

Please note that users can install browser extensions that controll the
internet proxy settings, and such an extension is displayed in the
internet settings page along with a disable button. Before this CL, the
disable button was not taking effect because the message handler that
disables the extension was not registered on that page.

Change-Id: I3b91fe9f9d5302a538def532d640700b21e7212c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4610362
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Jun Ishiguro <junis@google.com>
Cr-Commit-Position: refs/heads/main@{#1157963}