Clamp audio glitches and delay reported by OS.
We want to clamp the values so that if the OS calculates them
incorrectly, they do not add significant noise to the glitch and delay
stats.

It is important that the stats do not overflow when they are
accumulated in WebRtcAudioDeviceImpl. Given the bounds in this CL, we
can calculate how long it would take to reach overflow in the absolute
worst case.

Delay: 2^63 / (24*3600*192000*10*(10^6)) = 55 days
Glitch duration: 2^63 / (24*3600*100*1*(10^6)) = 1067519 days

Bug: 1449671
Change-Id: I35cfaec78bbedb2b70882f63aab50593752a540c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4607216
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Fredrik Hernqvist <fhernqvist@google.com>
Cr-Commit-Position: refs/heads/main@{#1158008}