[ios] Fix SetUpList signin crash when signin disabled
This change does three things:
 * fix a DCHECK crash in TwoScreensSigninCoordinator when signin is
   disabled by policy
 * change SetUpList to not include the signin item at all when signin is
   disabled when SetUpList loads
 * observe changes to auth service status: if signin is disabled while
   the SetUpList is loaded, it will be marked complete so that the user
   cannot attempt to sign in

The crash was because SceneController was calling stop on the signin
coordinator before it was started, and TwoScreensCoordinator has a
DCHECK to ensure that the screenProvider gets cleared. The fix for this
is to only init/alloc the screenProvider on start.

Fixed: 1454526
Change-Id: I6cf913baa69a5ced8ec319aef908c843e3362667
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4610445
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1158085}