iwa: Move some tests from the install command to the install helper
This CL moves tests that are testing the install command helper's
internals into a new unittest file for the helper, while leaving tests
that test the install command itself as is. It also adds some new tests
for the command helper.

In a follow-up CL, we could look into refactoring the install command's
tests to take advantage of possibly mocking the install command helper
instead of the DataRetriever.

Bug: 1444407
Change-Id: Iaf23e6f89509b050524fe6dae9a7b6a1cdbd36cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4593352
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Reviewed-by: Sonja Laurila <laurila@google.com>
Cr-Commit-Position: refs/heads/main@{#1158089}