WinMFVideoCapture: Hold onto single IMFSampleBuffer when posting from CaptureEngine callback thread to main capture thread
If texture processing for some reason takes too much time, there will be a
growing queue of posted OnIncomingCapturedDataInternal() tasks, each holding to
an IMFSampleBuffer.

Eventually, that may lead to E_MF_SAMPLEALLOCATOR error, because too many samples
are on the fly.

This CL makes sure that only one task will be queued and its arguments are
updated with the latest IMFSampleBuffer, if a new one arrives before the task
executes.

Note: the discarded samples are not reported as dropped frames.

Bug: 1120900, 1430048
Change-Id: I1cc4239c9125d0e3b07e7a3587112219d254396e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4614424
Commit-Queue: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158091}