lts: Return stable channel for ltc and lts build
ltc and lts builds have previously always been stable builds, but
starting with M120 that will no longer be the case. Images will then
contain ltc-channel and lts-channel int /etc/lsb-release. Keep
reporting the channel as stable in Chrome.

In almost all components we want to continue to treat ltc and lts
equivalent to stable. This CL maintains the continuity with the previous
implementation and avoids potential pitfalls.

Later CLs will make it possible to identify if we are actually on the
ltc or lts channel similarly to extended stable. This will be used for
example for UMA metrics.

Bug: b:284256808, b:285578735
Change-Id: I905a6aa4764ab2d3adf0d45fd6bab6ac003f72c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4614969
Reviewed-by: Roland Bock <rbock@google.com>
Commit-Queue: Victor-Gabriel Savu <vsavu@google.com>
Cr-Commit-Position: refs/heads/main@{#1158093}