Sort Blink features header, part 1/x.
This partially sorts the base::Features/base::FeatureParams. Part 2 will
sort the remainder of the header, and part 3 and onwards will sort the
.cc file.

This allows new base::Features to follow the sorting convention without
blocking on sorting the rest of this (very disorganized) file.

Bug: 1455020
Change-Id: I87949837757cdad1aec9546302dfbc851dd9fb63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4616728
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158185}